refactor election generator to manifest
link in date library
