+++ data/election_manifest_jefferson_county.json
