+++ test/electionguard/mocks/election.h
@@ -10,9 +10,6 @@
 extern "C" {
 #endif
 
-EG_API eg_electionguard_status_t
-eg_test_election_mocks_get_simple_election_from_file(eg_election_manifest_t **out_handle);
-
 EG_API eg_electionguard_status_t eg_test_election_mocks_get_fake_ciphertext_election(
   eg_election_manifest_t *in_manifest, eg_element_mod_p_t *in_public_key,
   eg_internal_manifest_t **out_manifest, eg_ciphertext_election_context_t **out_context);
