+++ test/electionguard/test_ballot_compact.cpp
@@ -1,6 +1,6 @@
 #include "../../src/electionguard/log.hpp"
 #include "mocks/ballot.hpp"
-#include "mocks/election.hpp"
+#include "mocks/manifest.hpp"
 
 #include <doctest/doctest.h>
 #include <electionguard/ballot.hpp>
@@ -13,7 +13,7 @@ using namespace std;
 TEST_CASE("Can serialize CompactPlaintextBallot")
 {
     // Arrange
-    auto manifest = ElectionGenerator::getFakeManifest(TWO_MOD_Q());
+    auto manifest = ManifestGenerator::getFakeManifest(TWO_MOD_Q());
     auto plaintext = BallotGenerator::getFakeBallot(*manifest);
     auto compact = CompactPlaintextBallot::make(*plaintext);
     auto msgpack = compact->toMsgPack();
